move DateJsonAdapter.kt
